+++ experiments/columnar/lib/runtime/Hash.cpp
@@ -5,7 +5,7 @@
 
 namespace columnar::runtime {
 
-void Hash::hash(llvm::ArrayRef<std::uint64_t> value,
+void Hash::hash(llvm::ArrayRef<std::uint32_t> value,
                 llvm::ArrayRef<std::size_t> sel,
                 llvm::MutableArrayRef<std::uint64_t> result) {
   for (auto [i, idx] : llvm::enumerate(sel)) {
